fix lint 2
